treewide: switch internal sampling format from f32 to i16
This brings various advantages:
- 50% less memory consumption
- no more range checks; every value is valid (instead of checking against NAN and so on)
- integer calculations are quicker than f32 calculations on chip
- most audio sources will be in i16 format anyway, less conversions beforehand

We still need to convert the i16 samples to f32 for the lowpass filter, tho. However,
there it is sufficient to just cast them. We do not need division to bring them into
-1.0..1.0 range. The cast is just one instruction in memory [0]. It's cheap.

[0] https://rust.godbolt.org/z/37n3Gq81z

+++ src/stdlib/recording.rs
@@ -105,7 +105,7 @@ pub fn start_detector_thread(
     let stream = input_dev
         .build_input_stream(
             &input_config,
-            move |data: &[f32], _info| {
+            move |data: &[i16], _info| {
                 log::trace!(
                     "audio input callback: {} samples ({} ms, sampling rate = {sampling_rate})",
                     data.len(),
